Drop Qt4 support
The background for this diff is that we want to specify Qt as a
dependency in the package.xml. However, it's only possible to specify
either Qt4 or Qt5. Thus we drop the support for Qt4 so that we can
specify Qt5 as a dependency in package.xml.

ROS melodic (which is already two years old) uses Qt5, thus this diff
shouldn't be too problematic.